fix: 添加独立更多信息
