+++ src/renderer/components/Toolbar/SettingContent/More/index.module.scss
@@ -1,0 +1,3 @@
+.content {
+  display: block;
+}
